+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Short accounts make long friends".
 - Tests for "Brackets, ultimate edition".
 - Tests for "Remaining card".
+- Tests for "Surakarta".
 
 ## [1.13.0] - 2022-09-30
 ### Added